move settings related files in one folder
