+++ src/ui/settings/key-bindings/edit-icon.js
