+++ src/ui/settings/key-bindings/trash-icon.js
